requirements.txt
torch
sentence-transformers~=5.1.0
chromadb~=1.0.20
rank-bm25~=0.2.2
langchain~=0.3.27
python-docx~=1.2.0
jieba>=0.42.1
spacy
google-genai

# python -m spacy download en_core_web_sm
# python -m spacy download zh_core_web_sm
numpy~=1.26.4
google~=3.0.0
protobuf~=6.32.0
transformers~=4.56.0